atari_experiments/src/dfa/node2vec/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
asn1crypto=0.24.0=py36_0
boto=2.48.0=py36h6e4cd66_1
boto3=1.5.32=py36_0
botocore=1.8.46=py36_0
bz2file=0.98=py36_0
ca-certificates=2018.03.07=0
cairo=1.14.12=h7636065_2
certifi=2018.1.18=py36_0
cffi=1.11.5=py36h9745a5d_0
chardet=3.0.4=py36h0f667ec_1
cryptography=2.2.2=py36h14c3975_0
cycler=0.10.0=py36h93f1223_0
dbus=1.13.2=h714fa37_1
decorator=4.3.0=py36_0
docutils=0.14=py36hb0f60f5_0
expat=2.2.5=he0dffb1_0
fontconfig=2.12.6=h49f89f6_0
freetype=2.8.1=0
gensim=3.4.0=py36h3010b51_0
glib=2.56.1=h000015b_0
graphite2=1.3.11=hf63cedd_1
graphviz=2.38.0=7
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
harfbuzz=1.7.6=h5f0a787_1
icu=58.2=h9c2bf20_1
idna=2.6=py36h82fb2a8_1
intel-openmp=2018.0.0=8
jmespath=0.9.3=py36hd3948f9_0
jpeg=9b=h024ee3a_2
kiwisolver=1.0.1=py36h764f252_0
libedit=3.1=heed3624_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=7.2.0=hdf63c60_3
libgfortran-ng=7.2.0=hdf63c60_3
libiconv=1.15=h63c8f33_5
libpng=1.6.34=hb9fc6fc_0
libstdcxx-ng=7.2.0=hdf63c60_3
libtiff=4.0.9=h28f6b97_0
libtool=2.4.6=h544aabb_3
libxcb=1.13=h1bed415_1
libxml2=2.9.8=hf84eae3_0
matplotlib=2.2.2=py36h0e671d2_1
mkl=2018.0.2=1
mkl_fft=1.0.1=py36h3010b51_0
mkl_random=1.0.1=py36h629b387_0
ncurses=6.0=h9df7e31_2
networkx=2.1=py36_0
numpy=1.14.2=py36hdbf6ddf_1
openssl=1.0.2o=h20670df_0
pandas=0.22.0=py36hf484d3e_0
pango=1.40.11=h8191d47_0
pcre=8.42=h439df22_0
pip=9.0.3=py36_0
pixman=0.34.0=hceecf20_3
pycparser=2.18=py36hf9f622e_1
pygraphviz=1.3=py36hd547939_0
pyopenssl=17.5.0=py36h20ba746_0
pyparsing=2.2.0=py36hee85983_1
pyqt=5.9.2=py36h751905a_0
pysocks=1.6.8=py36_0
python=3.6.5=hc3d631a_0
python-dateutil=2.7.2=py36_0
python-graphviz=0.8.2=py36_0
pytz=2018.4=py36_0
qt=5.9.5=h7e424d6_0
readline=7.0=ha6073c6_4
requests=2.18.4=py36he2e5f8d_1
s3transfer=0.1.13=py36_0
scipy=1.0.1=py36hfc37229_0
setuptools=39.0.1=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.11.0=py36h372c433_1
smart_open=1.5.7=py36_0
sqlite=3.22.0=h1bed415_0
tk=8.6.7=hc745277_3
tornado=5.0.1=py36_1
urllib3=1.22=py36hbe7ace6_0
wheel=0.31.0=py36_0
xz=5.2.3=h55aa19d_2
zlib=1.2.11=ha838bed_2
